move bess connection and monitor thread to Session
Instead of having a single global thread that monitors
all ports added to a TGENCLI instance, use one thread
per port.

The TGENCLI instance still tracks the sessions (and there
is a minor bug that we will lose an existing session if
we overwrite it with one whose port string is the same;
this will leave a running thread).

This is all totally untested (quick hack sesssion...)